+++ src/test/groovy/com/autonomousapps/internal/android/AndroidGradlePluginFactorySpec.groovy
@@ -24,6 +24,7 @@ final class AndroidGradlePluginFactorySpec extends Specification {
     '4.0'           | AndroidGradlePlugin4_0
     '4.0.0'         | AndroidGradlePlugin4_0
     '4.0.0-rc01'    | AndroidGradlePlugin4_0
+    '4.0.1'         | AndroidGradlePlugin4_0
     '4.1'           | AndroidGradlePlugin4_1
     '4.1.0'         | AndroidGradlePlugin4_1
     '4.1.0-alpha09' | AndroidGradlePlugin4_1