+++ src/functionalTest/groovy/com/autonomousapps/jvm/SpringBootSpec.groovy
@@ -1,0 +1,31 @@
+package com.autonomousapps.jvm
+
+import com.autonomousapps.jvm.projects.SpringBootProject
+import org.gradle.util.GradleVersion
+import spock.lang.Unroll
+
+import static com.autonomousapps.utils.Runner.build
+import static com.google.common.truth.Truth.assertThat
+
+final class SpringBootSpec extends AbstractJvmSpec {
+
+  @Unroll
+  def "does not suggest api dependencies (#gradleVersion)"() {
+    given:
+    def project = new SpringBootProject()
+    gradleProject = project.gradleProject
+
+    when:
+    build(gradleVersion, gradleProject.rootDir, ':buildHealth')
+
+    then:
+    assertThat(actualAdvice()).containsExactlyElementsIn(project.expectedAdvice)
+
+    where: 'Spring Boot requires Gradle 6.3+'
+    gradleVersion << [
+      GradleVersion.version('6.3'),
+      GradleVersion.version('6.4.1'),
+      GradleVersion.version('6.5')
+    ]
+  }
+}