+++ src/functionalTest/groovy/com/autonomousapps/jvm/projects/SpringBootProject.groovy
@@ -1,0 +1,53 @@
+package com.autonomousapps.jvm.projects
+
+import com.autonomousapps.AbstractProject
+import com.autonomousapps.advice.Advice
+import com.autonomousapps.kit.GradleProject
+import com.autonomousapps.kit.Plugin
+import com.autonomousapps.kit.Source
+import com.autonomousapps.kit.SourceType
+
+import static com.autonomousapps.kit.Dependency.okHttp
+
+final class SpringBootProject extends AbstractProject {
+
+  final GradleProject gradleProject
+
+  SpringBootProject() {
+    this.gradleProject = build()
+  }
+
+  private GradleProject build() {
+    def builder = newGradleProjectBuilder()
+    builder.withSubproject('proj') { s ->
+      s.sources = sources
+      s.withBuildScript { bs ->
+        bs.plugins = [Plugin.springBootPlugin, Plugin.javaPlugin]
+        bs.dependencies = [okHttp("implementation")]
+      }
+    }
+
+    def project = builder.build()
+    project.writer().write()
+    return project
+  }
+
+  private List<Source> sources = [
+    new Source(
+      SourceType.JAVA, "Main", "com/example",
+      """\
+        package com.example;
+        
+        import okhttp3.OkHttpClient;
+
+        public class Main {
+          public OkHttpClient ok() {
+            return new OkHttpClient.Builder().build();
+          }
+        }
+      """.stripIndent()
+    )
+  ]
+
+  final List<Advice> expectedAdvice = []
+}